+++ LoopBar-widget/src/main/java/com/cleveroad/loopbar/adapter/IOperationItem.java
@@ -5,5 +5,4 @@ public interface IOperationItem {
     boolean isVisible();
 
     void setVisible(boolean isVisible);
-
 }