+++ LoopBar-widget/src/main/java/com/cleveroad/loopbar/adapter/SimpleCategoriesAdapter.java
@@ -1,12 +1,13 @@
 package com.cleveroad.loopbar.adapter;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.cleveroad.loopbar.R;