+++ LoopBar-widget/src/main/java/com/cleveroad/loopbar/widget/AbstractSpacesItemDecoration.java
@@ -1,6 +1,7 @@
 package com.cleveroad.loopbar.widget;
 
-import android.support.v7.widget.RecyclerView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 class AbstractSpacesItemDecoration extends RecyclerView.ItemDecoration {