+++ LoopBar-widget/src/main/java/com/cleveroad/loopbar/widget/NegativeMarginFixColorDrawable.java
@@ -2,6 +2,7 @@ package com.cleveroad.loopbar.widget;
 
 import android.graphics.Canvas;
 import android.graphics.Region;
+import android.os.Build;
 
 class NegativeMarginFixColorDrawable extends android.graphics.drawable.ColorDrawable {
 
@@ -11,7 +12,11 @@ class NegativeMarginFixColorDrawable extends android.graphics.drawable.ColorDraw
 
     @Override
     public void draw(Canvas canvas) {
-        canvas.clipRect(getBounds(), Region.Op.REPLACE);
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+            canvas.clipRect(getBounds());
+        } else {
+            canvas.clipRect(getBounds(), Region.Op.REPLACE);
+        }
         super.draw(canvas);
     }
 }