+++ LoopBar-widget/src/main/java/com/cleveroad/loopbar/widget/SpacesLeftItemDecoration.java
@@ -1,8 +1,10 @@
 package com.cleveroad.loopbar.widget;
 
 import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 
 class SpacesLeftItemDecoration extends AbstractSpacesItemDecoration {
 
@@ -11,9 +13,9 @@ class SpacesLeftItemDecoration extends AbstractSpacesItemDecoration {
     }
 
     @Override
-    public void getItemOffsets(Rect outRect, View view, RecyclerView parent, RecyclerView.State state) {
+    public void getItemOffsets(@NonNull Rect outRect, @NonNull View view,
+                               @NonNull RecyclerView parent, @NonNull RecyclerView.State state) {
         if (parent.getChildAdapterPosition(view) == 0)
             outRect.left = getSpace();
     }
-
 }