+++ sample/src/main/java/com/cleveroad/sample/IFragmentReplacer.java
@@ -1,6 +1,7 @@
 package com.cleveroad.sample;
 
-import android.support.v4.app.Fragment;
+
+import androidx.fragment.app.Fragment;
 
 public interface IFragmentReplacer {
     void replaceFragment(Fragment fragment);