+++ sample/src/main/java/com/cleveroad/sample/fragments/MenuLoopBarFragment.java
@@ -1,9 +1,10 @@
 package com.cleveroad.sample.fragments;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.view.View;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 
 import com.cleveroad.loopbar.widget.LoopBarView;
 import com.cleveroad.sample.R;